test: moved some tests
